+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.57
 * Fixed iOS HLS initialization issue.
 * Fixed issue where video plays after resume even if it's not visible.
+* Updated User-Agent picking for Android.
 
 ## 0.0.56
 * Fixed empty data source notification issue.